add travis CI
this adds a travis CI configuration file which tests on

 * python 2.6
 * python 2.7
 * python 3.2
 * python 3.4
 * python 3.5
 * python 3.6
 * python 3.7
 * pypy

Tests on pyp3 are currently disabled due to pipenv being incompatible
with pypy3 (fix has been already merged but no new pipenv release yet).

Tests on jython are not provided since it's not supported by travis out
of the box (I managed to add it but still ran in to this [1] error, so
stopped looking into it).

[1] https://github.com/msgpack/msgpack-python/issues/303